+++ src/ModbusRTUSlave.h
@@ -8,7 +8,7 @@
 
 class ModbusRTUSlave : public ModbusSlaveLogic {
   public:
-    ModbusRTUSlave(Stream& serial, int8_t dePin = -1, int8_t rePin = -1);
+    ModbusRTUSlave(Stream& serial, int dePin = -1, int rePin = -1);
     void setResponseDelay(unsigned long responseDelay);
     void begin(uint8_t localUnitId, unsigned long baud, uint32_t config = SERIAL_8N1);
     bool poll();
